+++ services/explorer-api/src/http-server/routes/controllers/index.ts
@@ -1,0 +1,5 @@
+export * from "./inofficial.js";
+export * from "./blocks.js";
+export * from "./tx-effects.js";
+export * from "./contracts.js";
+export * from "./stats.js";
