+++ services/explorer-api/src/http-server/routes/index.ts
@@ -1,40 +1,40 @@
 import { Router } from "express";
-import * as controller from "./controller.js";
-import { routes } from "./routes_and_validation.js";
+import * as controller from "./controllers/index.js";
+import { paths } from "./paths_and_validation.js";
 
 export const init = ({ router }: { router: Router }) => {
   router.get("/l2/index", controller.GET_ROUTES);
 
-  router.get(routes.latestHeight, controller.GET_LATEST_HEIGHT);
-  router.get(routes.latestBlock, controller.GET_LATEST_BLOCK);
-  router.get(routes.block, controller.GET_BLOCK);
-  router.get(routes.blocks, controller.GET_BLOCKS);
+  router.get(paths.latestHeight, controller.GET_LATEST_HEIGHT);
+  router.get(paths.latestBlock, controller.GET_LATEST_BLOCK);
+  router.get(paths.block, controller.GET_BLOCK);
+  router.get(paths.blocks, controller.GET_BLOCKS);
 
   router.get(
-    routes.txEffectsByBlockHeight,
+    paths.txEffectsByBlockHeight,
     controller.GET_L2_TX_EFFECTS_BY_BLOCK_HEIGHT
   );
   router.get(
-    routes.txEffectByBlockHeightAndIndex,
+    paths.txEffectByBlockHeightAndIndex,
     controller.GET_L2_TX_EFFECT_BY_BLOCK_HEIGHT_AND_INDEX
   );
-  router.get(routes.txEffectsByTxHash, controller.GET_L2_TX_EFFECT_BY_TX_HASH);
+  router.get(paths.txEffectsByTxHash, controller.GET_L2_TX_EFFECT_BY_TX_HASH);
 
   router.get(
-    routes.contractInstancesByBlockHash,
+    paths.contractInstancesByBlockHash,
     controller.GET_L2_CONTRACT_INSTANCES_BY_BLOCK_HASH
   );
-  router.get(routes.contractInstance, controller.GET_L2_CONTRACT_INSTANCE);
+  router.get(paths.contractInstance, controller.GET_L2_CONTRACT_INSTANCE);
 
-  router.get(routes.statsTotalTxEffects, controller.GET_STATS_TOTAL_TX_EFFECTS);
+  router.get(paths.statsTotalTxEffects, controller.GET_STATS_TOTAL_TX_EFFECTS);
   router.get(
-    routes.statsTotalTxEffectsLast24h,
+    paths.statsTotalTxEffectsLast24h,
     controller.GET_STATS_TOTAL_TX_EFFECTS_LAST_24H
   );
-  router.get(routes.statsTotalContracts, controller.GET_STATS_TOTAL_CONTRACTS);
-  router.get(routes.statsAverageFees, controller.GET_STATS_AVERAGE_FEES);
+  router.get(paths.statsTotalContracts, controller.GET_STATS_TOTAL_CONTRACTS);
+  router.get(paths.statsAverageFees, controller.GET_STATS_AVERAGE_FEES);
   router.get(
-    routes.statsAverageBlockTime,
+    paths.statsAverageBlockTime,
     controller.GET_STATS_AVERAGE_BLOCK_TIME
   );
